Update documentation for Copy and Zip classes
