--- src/main/java/com/example/FileCopier.java
@@ -1,5 +1,0 @@
-package com.example;
-
-public interface FileCopier {
-    
-}
