+++ src/main/java/com/example/FileCopy/FileCopier.java
@@ -1,0 +1,8 @@
+package com.example.FileCopy;
+
+/*
+ * The `FileCopier` interface provides a contract for copying files from one location to another.
+ */
+public interface FileCopier {
+    
+}
