--- src/main/java/com/example/Unzip.java
@@ -1,5 +1,0 @@
-package com.example;
-
-public interface Unzip {
-    public String unzip(String zipFilePath, String outputFolder);
-}
